code/appendix_hybrid_search/requirements.txt
# Core ML/Data
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# BM25 (fast sparse-matrix based)
bm25s[core]>=0.2.0

# SPLADE (learned sparse retrieval)
transformers>=4.30.0
torch>=2.0.0

# Dense retrieval (Sentence-BERT + FAISS)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Parquet support
pyarrow>=12.0.0

# Progress bars
tqdm>=4.60.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
